// src/utils/enums.ts
export enum PhoneAction {
  Answer = "ANSWER",
  Call = "CALL",
  HangUp = "HANG UP",
  Reject = "REJECT",
}
export enum Status {
  Idle = "IDLE",
  Ringing = "RINGING",
  Talking = "TALKING",
  PlacingCall = "PLACING CALL",
  RemoteIsRinging = "REMOTE IS RINGING",
  RemoteUnknown = "REMOTE UNKNOWN",
  RemoteBusy = "REMOTE BUSY",
  RemoteRejected = "REMOTE REJECTED",
}